+++ test/libsolidity/syntaxTests/lvalues/calldata_index_access.sol
@@ -1,0 +1,7 @@
+contract C {
+    function f(uint256[] calldata x) external pure {
+        x[0] = 42;
+    }
+}
+// ----
+// TypeError: (74-78): Calldata arrays are read-only.